+++ src/main/java/io/kahu/hawaii/util/call/sql/response/ListResponseHandler.java
@@ -37,9 +37,8 @@ public class ListResponseHandler<R extends ResultSet, T> implements ResponseHand
     @Override
     public void addToResponse(ResultSet resultSet, Response<List<T>> response) throws ServerException {
         try {
-            if (resultSet.isBeforeFirst()) {
-                // Cry bloody murder!
-            }
+            assert (resultSet.isBeforeFirst());
+            
             int i = 0;
             List<T> list = new ArrayList<>();
             while (resultSet.next()) {